ml-serving/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.3
pydantic-settings==2.6.1
torch==2.5.1
transformers==4.44.0
pillow==12.0.0
python-multipart==0.0.9
